pyproject.toml
[project]
name = "doc-flow-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Dmitry",email = "kde98@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiogram>=3.5.0",
  "SQLAlchemy>=2.0.0",
  "python-dotenv>=1.0.0",
  "httpx>=0.27.0",
  "pydantic>=2.5.0",
  "docxtpl>=0.16.7",
  "loguru>=0.7.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
